Move installed binaries out of $CARGO_HOME
The intent is to make the container workable with $CARGO_HOME mounted
from the host, to preserve Cargo caches and the authentication token
between container runs.

If a container run invokes rustup, this may override the binaries
moved to /usr/local/bin with freshly installed binaries in $CARGO_HOME/bin/.
This is perfectly fine for incremental container images.