Add Debian 12 "Bookworm" variant
In addition to the current stable "Bullseye" and oldstable "Buster", it may
be a good time to start publishing images for the upcoming Debian "Bookworm"
release. It has gone into the first stage of freezing before its upcoming
launch later this year, so it's a good time for people to start testing
their apps on it, as it is both reasonably settled/stable and also fairly
imminent.