Add `musl-dev`
`gcc` is already a package installed, but it does not install a `libc` candidate with it. `musl-dev` contains the necessary development files (`crti.o`, `crtn.o`, `Scrt1.o`, `libc.so`, etc.).

This was prior requested in https://github.com/rust-lang/docker-rust/issues/68 and is necessary for `proc-macro`s/build scripts to compile, without additional configuration/setup. It also is apparently just a 2% size increase which many consumers are already incurring due to how only effectively bare-metal compilation is possible without it.